requirement.txt
langchain>=0.1.13
langchain-google-genai>=0.0.7
google-generativeai>=0.3.2
python-dotenv>=1.0.1
matplotlib>=3.7.0
wordcloud>=1.8.2.2
streamlit>=1.32.0
